requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
casbin==1.33.0
python-jose==3.3.0
passlib==1.7.4
pydantic==2.5.2
python-multipart==0.0.6
bcrypt==4.0.1
email-validator==2.1.0
pytest==7.4.3
httpx==0.25.2